+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 #### Changes
+* Node: Added FUNCTION STATS commands ([#2081](https://github.com/valkey-io/valkey-glide/pull/2081))
 * Node: Added SORT commands ([#2028](https://github.com/valkey-io/valkey-glide/pull/2028))
 * Node: Added LASTSAVE command ([#2059](https://github.com/valkey-io/valkey-glide/pull/2059))
 * Node: Added LCS command ([#2049](https://github.com/valkey-io/valkey-glide/pull/2049))